Updating to release version 0.1.1 which includes the following:
- Fix bug in create embeding response
- Update client to support multiple deployments at run time so we can use gpt and ada at the same time with a single client
- Revert to java 11 for more compatibility with existing projects

+++ src/main/java/ca/int13/azureai/api/classes/AzureAICreateEmbedingResponseData.java
@@ -31,7 +31,7 @@ import java.util.ArrayList;
 public class AzureAICreateEmbedingResponseData {
     private String object;
     private int index;
-    private ArrayList<Long> embedding;
+    private ArrayList<Float> embedding;
 
     /**
      * @return the object
@@ -64,14 +64,14 @@ public class AzureAICreateEmbedingResponseData {
     /**
      * @return the embedding
      */
-    public ArrayList<Long> getEmbedding() {
+    public ArrayList<Float> getEmbedding() {
         return embedding;
     }
 
     /**
      * @param embedding the embedding to set
      */
-    public void setEmbedding(ArrayList<Long> embedding) {
+    public void setEmbedding(ArrayList<Float> embedding) {
         this.embedding = embedding;
     }
